+++ src/config-type.ts
@@ -55,7 +55,10 @@ export interface AppConfig {
   app: Express; // or your custom express app
 }
 
-type InputSource = keyof Pick<Request, "query" | "body" | "files" | "params">;
+export type InputSource = keyof Pick<
+  Request,
+  "query" | "body" | "files" | "params"
+>;
 export type InputSources = Record<Method, InputSource[]>;
 
 type Headers = Record<string, string>;
@@ -86,7 +89,7 @@ export interface CommonConfig<TAG extends string = string> {
   // what request properties are combined into input for endpoints and middlewares
   // the order matters: priority from lowest to highest
   // default: { get: [query, params], post: [body, params, files],
-  // put: [body, params], patch: [body, params], delete: [body, query, params] }
+  // put: [body, params], patch: [body, params], delete: [query, params] }
   inputSources?: Partial<InputSources>;
   // optional endpoints tagging configuration, example: { users: "Everything about the users" }
   tags?: TagsConfig<TAG>;