+++ tests/unit/endpoint.spec.ts
@@ -9,7 +9,6 @@ import {
 } from "../../src";
 import { Endpoint } from "../../src/endpoint";
 import { IOSchemaError } from "../../src/errors";
-import { mimeJson } from "../../src/mime";
 import { serializeSchemaForTest } from "../helpers";
 
 describe("Endpoint", () => {
@@ -18,14 +17,13 @@ describe("Endpoint", () => {
       const endpointMock = new Endpoint({
         methods: ["get", "post", "put", "delete", "patch"],
         inputSchema: z.object({}),
-        mimeTypes: [mimeJson],
         outputSchema: z.object({}),
         handler: jest.fn(),
-        resultHandler: {
-          getPositiveResponse: jest.fn(),
-          getNegativeResponse: jest.fn(),
+        resultHandler: createResultHandler({
+          getPositiveResponse: () => z.string(),
+          getNegativeResponse: () => z.string(),
           handler: jest.fn(),
-        },
+        }),
         middlewares: [],
       });
       expect(endpointMock.getMethods()).toEqual([
@@ -41,14 +39,13 @@ describe("Endpoint", () => {
       const endpointMock = new Endpoint({
         method: "patch",
         inputSchema: z.object({}),
-        mimeTypes: [mimeJson],
         outputSchema: z.object({}),
         handler: jest.fn(),
-        resultHandler: {
-          getPositiveResponse: jest.fn(),
-          getNegativeResponse: jest.fn(),
+        resultHandler: createResultHandler({
+          getPositiveResponse: () => z.string(),
+          getNegativeResponse: () => z.string(),
           handler: jest.fn(),
-        },
+        }),
         middlewares: [],
       });
       expect(endpointMock.getMethods()).toEqual(["patch"]);
@@ -310,23 +307,24 @@ describe("Endpoint", () => {
         output: z.object({}),
         handler: jest.fn(),
       });
-      expect(endpoint.getInputSchema()).toEqual(input);
+      expect(endpoint.getSchema("input")).toEqual(input);
     });
   });
 
-  describe(".getOutputSchema()", () => {
-    test("should return output schema", () => {
-      const factory = new EndpointsFactory(defaultResultHandler);
-      const output = z.object({
+  describe(".outputSchema", () => {
+    test("should be the output schema", () => {
+      const outputSchema = z.object({
         something: z.number(),
       });
-      const endpoint = factory.build({
+      const endpoint = new Endpoint({
         method: "get",
-        input: z.object({}),
-        output,
+        middlewares: [],
+        inputSchema: z.object({}),
+        outputSchema,
         handler: jest.fn(),
+        resultHandler: defaultResultHandler,
       });
-      expect(endpoint.getOutputSchema()).toEqual(output);
+      expect(endpoint.getSchema("output")).toEqual(outputSchema);
     });
   });
 
@@ -343,7 +341,7 @@ describe("Endpoint", () => {
         handler: jest.fn(),
       });
       expect(
-        serializeSchemaForTest(endpoint.getPositiveResponseSchema())
+        serializeSchemaForTest(endpoint.getSchema("positive"))
       ).toMatchSnapshot();
     });
   });
@@ -361,7 +359,7 @@ describe("Endpoint", () => {
         handler: jest.fn(),
       });
       expect(
-        serializeSchemaForTest(endpoint.getNegativeResponseSchema())
+        serializeSchemaForTest(endpoint.getSchema("negative"))
       ).toMatchSnapshot();
     });
   });
@@ -375,7 +373,7 @@ describe("Endpoint", () => {
         output: z.object({}),
         handler: jest.fn(),
       });
-      expect(endpoint.getPositiveMimeTypes()).toEqual(["application/json"]);
+      expect(endpoint.getMimeTypes("positive")).toEqual(["application/json"]);
     });
   });
 
@@ -388,7 +386,7 @@ describe("Endpoint", () => {
         output: z.object({}),
         handler: jest.fn(),
       });
-      expect(endpoint.getNegativeMimeTypes()).toEqual(["application/json"]);
+      expect(endpoint.getMimeTypes("negative")).toEqual(["application/json"]);
     });
   });
 
@@ -717,7 +715,6 @@ describe("Endpoint", () => {
           new Endpoint({
             method: "get",
             inputSchema: z.object({}).transform(() => []),
-            mimeTypes: [mimeJson],
             outputSchema: z.object({}),
             handler: jest.fn(),
             resultHandler: {
@@ -737,7 +734,6 @@ describe("Endpoint", () => {
           new Endpoint({
             method: "get",
             inputSchema: z.object({}),
-            mimeTypes: [mimeJson],
             outputSchema: z.object({}).transform(() => []),
             handler: jest.fn(),
             resultHandler: {