+++ tests/unit/result-handler.spec.ts
@@ -1,5 +1,11 @@
 import { Request, Response } from "express";
-import { createHttpError, defaultResultHandler, withMeta, z } from "../../src";
+import {
+  InputValidationError,
+  createHttpError,
+  defaultResultHandler,
+  withMeta,
+  z,
+} from "../../src";
 import { metaProp } from "../../src/metadata";
 
 let loggerMock: any;
@@ -61,15 +67,17 @@ describe("ResultHandler", () => {
         url: "http://something/v1/anything",
       };
       defaultResultHandler.handler({
-        error: new z.ZodError([
-          {
-            code: "invalid_type",
-            message: "Expected string, got number",
-            path: ["something"],
-            expected: "string",
-            received: "number",
-          },
-        ]),
+        error: new InputValidationError(
+          new z.ZodError([
+            {
+              code: "invalid_type",
+              message: "Expected string, got number",
+              path: ["something"],
+              expected: "string",
+              received: "number",
+            },
+          ])
+        ),
         input: { something: 453 },
         output: { anything: 118 },
         request: requestMock as Request,