+++ CHANGELOG.md
@@ -1,5 +1,84 @@
 # Changelog
 
+## Version 9
+
+### v9.0.0-beta4
+
+- This release contains the feature from version [8.11.0](#v8110).
+- **BREAKING** changes:
+  - `createApiResponse()` method is removed. Read the release notes to v8.11.0 for migration strategy.
+- Potentially **BREAKING** changes:
+  - The following changes correspond to the entities that are not supposed to be used directly, however they are public.
+  - `Endpoint::constructor()`
+    - `mimeTypes` property is removed from the argument.
+  - `Endpoint` public methods replaced:
+    - `getPositiveStatusCode()` —> `getStatusCode("positive")`
+    - `getNegativeStatusCode()` —> `getStatusCode("negative")`
+    - `getInputSchema()` —> `getSchema("input")`
+    - `getOutputSchema()` —> `getSchema("output")`
+    - `getPositiveResponseSchema()` —> `getSchema("positive")`
+    - `getNegativeResponseSchema()` —> `getSchema("negative")`
+    - `getInputMimeTypes()` —> `getMimeTypes("input")`
+    - `getPositiveMimeTypes()` —> `getMimeTypes("positive")`
+    - `getNegativeMimeTypes()` —> `getMimeTypes("negative")`
+
+### v9.0.0-beta3
+
+- This release contains the feature from version [8.10.0](#v8100).
+
+### v9.0.0-beta2
+
+- Potentially **BREAKING** changes:
+  - Fixed problem #787, reported and resolved by [@TheWisestOne](https://github.com/TheWisestOne).
+    - Validation errors thrown from within the Middlewares and Endpoint handlers unrelated to the IO do now lead to the
+      status code `500` instead of `400`, when you're using the `defaultResultHandler` or `defaultEndpointsFactory`.
+      - It enables you to use zod (via the exposed `z` namespace) for the internal needs of your implementation, such as
+        validating the data coming from your database, for example.
+    - Historically, `ZodError` meant the error related to the input validation, but it's changed.
+      - New error class created: `InputValidationError`.
+      - If you have a custom `ResultHandler` that relies on `ZodError` for responding with `400` code, you need to
+        change that condition to `InputValidationError` in order to keep that behaviour.
+    - Luckily, the following entities were exposed and became available for the convenience of your migration:
+      - `OutputValidationError`,
+      - `InputValidationError` _(new)_,
+      - `getMessageFromError()`,
+      - `getStatusCodeFromError()`.
+    - Consider using `getStatusCodeFromError()` inside your custom `ResultHandler`, or make the following changes:
+
+```typescript
+// Your custom ResultHandler
+// Before: if you're having an expression like this:
+if (error instanceof z.ZodError) {
+  response.status(400);
+}
+// After: replace it to this:
+if (error instanceof InputValidationError) {
+  response.status(400);
+}
+// Or: consider the alternative:
+const statusCode = getStatusCodeFromError(error);
+const message = getMessageFromError(error);
+response.status(statusCode);
+```
+
+### v9.0.0-beta1
+
+- This release is based on version 8.9.4.
+- Potentially **BREAKING** changes:
+  - Fixed issue #820, reported and resolved by [@McMerph](https://github.com/McMerph).
+    - Request `body` is no longer considered as an input source for `DELETE` request.
+    - Despite the fact that this method MAY contain `body` (it's not explicitly prohibited), it's currently considered
+      a bad practice to rely on it. Also, it led to a syntax error in the generated documentation according to OpenAPI
+      3.0 specification.
+    - In case you have such Endpoints that rely on inputs collected from `DELETE` request body and want to continue,
+      add the following property to your configuration in order to keep the previous behavior without changes to your
+      implementation.
+    - Read the [customization instructions](https://github.com/RobinTail/express-zod-api#customizing-input-sources).
+
+```yaml
+inputSources: { delete: ["body", "query", "params"] }
+```
+
 ## Version 8
 
 ### v8.11.0